Fixed player movement.
Keyboard states were originally stored using SDL_GetKeyboardState, but this was resulting in code for other movement keys being called as well as the button that was currently pressed. I changed it to use SDL_KEYUP/SDL_KEYDOWN events instead. The player now moves smoothly when holding the WASD buttons down, however it's not locked to a specific framerate. Also if I try to calculate the timeStep variable in Player::Update(), the back and right keys don't move the player.