Sliding collision
The player now slides along the edges of objects when colliding with them, rather than just stopping.
The Collision Manager first tests if a collision is taking place. It then tests if the collision is coming from the side or from the top or bottom. It then resets the player position along the appropriate axis.