GLM and collision code
Added GLM and replaced SDL_Rect with glm::vec2 where the width and height wasn't needed, mainly in the Player class.

Removed CollisionManager and moved all the collision code into Player temporarily.

Moved the Camera object out of ControllerGame and into Scene. The player is still kept in ControllerGame, but it's managed as a reference in Scene.